pyproject.toml
[project]
name = "ai-governance-register"
version = "0.1.0"
description = "AI Compliance & Governance Register — internal AI system registry and governance tracker"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.39.0",
    "reportlab>=4.0.0",
    "python-docx>=1.0.0",
    "requests>=2.31.0",
    "streamlit>=1.28.0",
    "plotly>=5.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
